fix: :sparkles: Added report menus, also added titles for sales and purchases.
In the file located in modules/aggCompra.py the input has been modified to give an example of how to enter the date and to add the global clear screen function.   In the modules/aggTransaction file as well as in the purchase file the example input was added.  In addition in the file reports the error that arose when executing the option that starts the function to generate reports was corrected.

@By ayelmercorzo